refactor: update rule module
